requirements.txt
openai
langchain
langgraph
faiss-cpu
fastapi
uvicorn
pydantic
pandas
streamlit
python-dotenv
langchain_openai
langchain_community